minor improvements based on pr feedback
Pretty print json output
Move fileFormat to common api config
Adjust default file extension when using json file format
