+++ openapi-maven-plugin/src/main/java/io/github/kbuntrock/configuration/ApiConfiguration.java
@@ -8,16 +8,14 @@ import org.apache.maven.plugins.annotations.Parameter;
 
 public class ApiConfiguration extends CommonApiConfiguration {
 
+	private static final String DEFAULT_FILENAME = "spec-open-api.yml";
 	/**
 	 * A list of location to find api endpoints. A location could be a class or a package
 	 */
 	@Parameter(required = true)
 	private List<String> locations;
 	@Parameter
-	private String filename = "spec-open-api.yml";
-
-	@Parameter(defaultValue = "yaml")
-	private String fileFormat;
+	private String filename = DEFAULT_FILENAME;
 
 	protected String baseFreeField;
 	@Parameter
@@ -39,14 +37,6 @@ public class ApiConfiguration extends CommonApiConfiguration {
 
 	public void setFilename(final String filename) {
 		this.filename = filename;
-	}
-
-	public String getFileFormat() {
-		return fileFormat;
-	}
-
-	public void setFileFormat(String fileFormat) {
-		this.fileFormat = fileFormat;
 	}
 
 	public OperationIdHelper getOperationIdHelper() {
@@ -86,6 +76,7 @@ public class ApiConfiguration extends CommonApiConfiguration {
 		merged.defaultProduceConsumeGuessing = copy.defaultProduceConsumeGuessing;
 		merged.pathEnhancement = copy.pathEnhancement;
 		merged.pathPrefix = copy.pathPrefix;
+		merged.fileFormat = copy.fileFormat;
 		merged.loopbackOperationName = copy.loopbackOperationName;
 		merged.operationId = copy.operationId;
 		merged.freeFields = copy.freeFields;
@@ -107,7 +98,6 @@ public class ApiConfiguration extends CommonApiConfiguration {
 
 		merged.setLocations(locations);
 		merged.setFilename(filename);
-		merged.setFileFormat(fileFormat);
 		merged.setMergeFreeFields(mergeFreeFields);
 
 		if(!tag.getSubstitutions().isEmpty()) {
@@ -130,6 +120,12 @@ public class ApiConfiguration extends CommonApiConfiguration {
 		}
 		if(pathPrefix != null) {
 			merged.setPathPrefix(pathPrefix);
+		}
+		if(fileFormat != null) {
+			merged.setFileFormat(fileFormat);
+		}
+		if("json".equals(fileFormat) && DEFAULT_FILENAME.equals(filename)){
+			merged.filename = DEFAULT_FILENAME.replace(".yml", ".json");
 		}
 		if(loopbackOperationName != null) {
 			merged.setLoopbackOperationName(loopbackOperationName);
